# 8359218: RISC-V: Only enable CRC32 intrinsic when AvoidUnalignedAccess == false